MouseInputExample.kt: Extract movement line restriction in own class
